Add path comments and docstrings
