+++ f2/apps/douyin/crawler.py
@@ -87,7 +87,8 @@ from f2.apps.douyin.utils import (
     TokenManager,
     XBogusManager,
 )
-from f2.crawlers.base_crawler import BaseCrawler, WebSocketCrawler
+from f2.crawlers.base_crawler import BaseCrawler
+from f2.crawlers.websocket_crawler import WebSocketCrawler
 from f2.i18n.translator import _
 from f2.log.logger import logger, trace_logger
 from f2.utils.http.endpoint import BaseEndpointManager
