+++ f2/apps/tiktok/crawler.py
@@ -50,7 +50,8 @@ from f2.apps.tiktok.proto.tiktok_webcast_pb2 import (
     UserFanTicket,
 )
 from f2.apps.tiktok.utils import ClientConfManager, XBogusManager
-from f2.crawlers.base_crawler import BaseCrawler, WebSocketCrawler
+from f2.crawlers.base_crawler import BaseCrawler
+from f2.crawlers.websocket_crawler import WebSocketCrawler
 from f2.i18n.translator import _
 from f2.log.logger import logger, trace_logger
 from f2.utils.http.endpoint import BaseEndpointManager
