+++ f2/dl/base_downloader.py
@@ -4,14 +4,15 @@ import asyncio
 import sys
 import traceback
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set, Union
+from typing import Any, Dict, List, Optional, Union
 
-import aiofiles
+import aiofiles  # type: ignore
 import httpx
 from rich.progress import TaskID
 
 from f2.cli.cli_console import RichConsoleManager
 from f2.crawlers.base_crawler import BaseCrawler
+from f2.dl.m3u8 import M3U8DownloadMixin
 from f2.exceptions.api_exceptions import APIRetryExhaustedError
 from f2.i18n.translator import _
 from f2.log.logger import logger, trace_logger
@@ -20,16 +21,11 @@ from f2.utils.file.path import ensure_path
 from f2.utils.http.utils import (
     get_chunk_size,
     get_content_length,
-    get_segments_from_m3u8,
     trim_filename,
 )
 
-# 最大片段缓存数量，超过这个数量就会进行清理
-# (Maximum segment cache count, clear when it exceeds this count)
-MAX_SEGMENT_COUNT = 1000
 
-
-class BaseDownloader(BaseCrawler):
+class BaseDownloader(M3U8DownloadMixin, BaseCrawler):
     """
     基础下载器 (Base Downloader)
 
@@ -340,174 +336,6 @@ class BaseDownloader(BaseCrawler):
             visible=False,
         )
         logger.debug(_("文件已保存到：{0}").format(full_path))
-
-    async def download_m3u8_stream(
-        self,
-        task_id: TaskID,
-        url: str,
-        full_path: Union[str, Path],
-    ) -> None:
-        """
-        下载m3u8流视频 (Download m3u8 stream video)
-
-        Args:
-            task_id (TaskID): 任务ID (Task ID)
-            url (str): m3u8文件的URL (m3u8 file URL)
-            full_path (Union[str, Path]): 保存路径 (Save path)
-
-        Note:
-            由于直播流的特殊性，可能会出现直播结束、账号在别处进入直播间等情况，导致直播流无法下载。
-
-            直播流的大小不确定，因此无法准确计算下载进度，只能根据下载的块大小来更新进度条。
-
-            可能会出现 httpx.RemoteProtocolError 错误，这是由于服务器返回的块大小未严格遵守 HTTP 规范。
-            非代码问题，而是服务器问题，跳过该片段处理。
-            Issues: https://github.com/encode/httpx/issues/1927
-        """
-        async with self.semaphore:
-            full_path = self._ensure_path(full_path)
-            # 设置默认下载总量 (Set default total download)
-            total_downloaded = 10240000
-            # 默认块大小 (Default chunk size)
-            default_chunks = 409600
-            # 记录已经下载的片段序号
-            # (Record the segment number that has been downloaded)
-            downloaded_segments: Set = set()
-
-            while not SignalManager.is_shutdown_signaled():
-                try:
-                    segments = await get_segments_from_m3u8(url)
-
-                    if not segments:
-                        logger.debug(_("m3u8片段为空，直播流已结束"))
-                        logger.info(
-                            _("[green][  完成  ]：{0}[/green]").format(
-                                Path(full_path).name
-                            )
-                        )
-                        await self.progress.update(
-                            task_id,
-                            description=_("[green][  完成  ]：[/green]"),
-                            filename=trim_filename(full_path.name, 45),
-                            state="completed",
-                            visible=False,
-                        )
-                        return
-
-                    # 确保目标路径存在 (Ensure target path exists)
-                    full_path.parent.mkdir(parents=True, exist_ok=True)
-
-                    async with aiofiles.open(full_path, "ab") as file:
-                        for segment in segments:
-                            if SignalManager.is_shutdown_signaled():
-                                break
-
-                            # 检查是否已经下载过该片段 (Check if the segment has been downloaded)
-                            if segment.absolute_uri not in downloaded_segments:
-                                ts_url = segment.absolute_uri
-                                ts_content_length = await get_content_length(
-                                    ts_url,
-                                    self.headers,
-                                    self.proxies,
-                                )
-                                if ts_content_length == 0:
-                                    ts_content_length = default_chunks
-                                    logger.debug(
-                                        _(
-                                            "无法读取该TS文件字节长度，将使用默认400kb块大小处理数据"
-                                        )
-                                    )
-
-                                try:
-                                    ts_request = self.aclient.build_request(
-                                        "GET", ts_url, headers=self.headers
-                                    )
-                                    ts_response = await self.aclient.send(
-                                        ts_request, stream=True
-                                    )
-
-                                    async for chunk in ts_response.aiter_bytes(
-                                        get_chunk_size(ts_content_length)
-                                    ):
-                                        if SignalManager.is_shutdown_signaled():
-                                            break
-
-                                        # 直播流分块下载，每次下载后更新进度条
-                                        # (Live stream block download, update progress bar after each download)
-                                        await file.write(chunk)
-                                        total_downloaded += len(chunk)
-                                        await self.progress.update(
-                                            task_id,
-                                            advance=len(chunk),
-                                            total=total_downloaded,
-                                        )
-
-                                    # 记录已经下载的片段序号
-                                    # (Record the segment number that has been downloaded)
-                                    downloaded_segments.add(segment.absolute_uri)
-
-                                except httpx.ReadTimeout:
-                                    logger.warning(_("下载超时：跳过该 TS 文件片段"))
-                                    continue
-
-                                except httpx.RemoteProtocolError as e:
-                                    logger.error(
-                                        _(
-                                            "服务器返回的块大小未严格遵守 HTTP 规范，跳过该片段。错误信息：{0}"
-                                        ).format(e)
-                                    )
-                                    continue
-
-                                finally:
-                                    await ts_response.aclose()
-                            else:
-                                logger.debug(
-                                    _("跳过已下载的片段，URL：{0}").format(
-                                        segment.absolute_uri
-                                    )
-                                )
-
-                            # 每下载一定数量的片段后，清理一次集合
-                            # (After downloading a certain number of segments, clean up the collection)
-                            if len(downloaded_segments) > MAX_SEGMENT_COUNT:
-                                downloaded_segments = set()
-
-                    # 等待片段时长，避免过快下载
-                    # (Wait for the segment duration to avoid downloading too fast)
-                    await asyncio.sleep(segment.duration)
-
-                except httpx.HTTPStatusError as e:
-                    if e.response.status_code == 404:
-                        logger.debug(_("m3u8文件或ts文件未找到，当前直播已结束"))
-                    elif e.response.status_code == 504:
-                        logger.warning(_("[red]网关超时，无法下载直播流[/red]"))
-                    else:
-                        logger.debug(_("HTTP错误：{0}").format(e))
-                        logger.error(_("[red]m3u8文件下载失败，但文件已保存[/red]"))
-
-                    logger.info(
-                        _("[green][  完成  ]：{0}[/green]").format(Path(full_path).name)
-                    )
-                    await self.progress.update(
-                        task_id,
-                        description=_("[red][  完成  ]：[/red]"),
-                        filename=trim_filename(full_path.name, 45),
-                        state="completed",
-                        visible=False,
-                    )
-                    logger.debug(_("直播流文件已保存到：{0}").format(full_path))
-                    return
-
-                except Exception as e:
-                    trace_logger.error(traceback.format_exc())
-                    logger.error(_("m3u8文件解析失败：{0}").format(e))
-                    await self.progress.update(
-                        task_id,
-                        description=_("[red][  失败  ]：[/red]"),
-                        filename=trim_filename(full_path.name, 45),
-                        state="completed",
-                    )
-                    return
 
     async def initiate_download(
         self,
